backend/requirements.txt
flask
flask-cors
pillow
pyzbar
opencv-python-headless
numpy
werkzeug